+++ packages/form/src/form/model.ts
@@ -1,8 +1,10 @@
 import type { Ref } from '@/lib/svelte.svelte.js';
+import type { KeyedArray2 } from "@/lib/keyed-array.svelte.js";
 import {
   getSchemaDefinitionByPath,
   type Path,
   type Schema,
+  type SchemaArrayValue,
   type SchemaValue,
 } from "@/core/index.js";
 
@@ -13,6 +15,9 @@ export type FormValue = SchemaValue | undefined;
 // TODO: Remove in v4
 /** @deprecated use `Ref` from `lib/types` */
 export type ValueRef<T> = Ref<T>
+export type KeyedFieldValues = KeyedArray2<number, FieldValue>
+
+export type KeyedArraysMap = WeakMap<SchemaArrayValue, KeyedFieldValues>;
 
 export const DEFAULT_BOOLEAN_ENUM = [true, false];