+++ packages/testing/CHANGELOG.md
@@ -1,5 +1,13 @@
 # testing
 
+## 1.0.11
+
+### Patch Changes
+
+- Updated dependencies [[`8772516`](https://github.com/x0k/svelte-jsonschema-form/commit/877251689dc60c91b7d2a257c3f05665d9a7ea96), [`66ca1a5`](https://github.com/x0k/svelte-jsonschema-form/commit/66ca1a5001a65fa794e8fcb6b1362c8365ff6686), [`123fc70`](https://github.com/x0k/svelte-jsonschema-form/commit/123fc70c1e7ef7ad1d045584790d48c957b6f03c), [`5f64b23`](https://github.com/x0k/svelte-jsonschema-form/commit/5f64b23a9baedba6d6df2c4933acb180a6300668), [`2c8501b`](https://github.com/x0k/svelte-jsonschema-form/commit/2c8501b59a7beef6ac20d88a2f16bdc350779183), [`b171e71`](https://github.com/x0k/svelte-jsonschema-form/commit/b171e7116617eb8d47f7f9552c5cb5235ca1a86e)]:
+  - @sjsf/form@2.3.0
+  - @sjsf/ajv8-validator@2.3.0
+
 ## 1.0.10
 
 ### Patch Changes